REDO
After pairing, make the room, send out the room id and if the user got the room id the UI will trigger a checking when the chat view is loaded, checking the roomid is valid and the user with its own user id added to that room. The room on the server not representing the current state, it contains the users by id to be able to check.